Added support for crypt and plaintext authentication. Modified tests and code to fully support ruby 1.8.7, too. Tests can now be run via rake, as well.
Update readme to reflection that authentication works now.

Renaming to get the markdown goodness.

Adding rbx-require-relative to get the desired behaviors for require_relative when running on ruby 1.8.7. This ruby version is listed a supported platform in .travis.yml.

Making both gem and tests happy on Ruby 1.8.7.

Throwing in some tests that should allow testing of plain and crypt auth scenarios.

Key did not get parsed on ruby 1.8.x because ruby 1.8.x is stupid. But sure, preserving compatibility.